requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
jinja2>=3.1
requests>=2.31
SQLAlchemy>=2.0
alembic>=1.13
python-dotenv>=1.0
python-jose[cryptography]>=3.3
openai>=1.0
